Don't add "autogenerate" comment by default
Having such a comment in the final docs HTML might be undesirable, and this is
the case e.g. when the markdown is rendered with Hugo. Additionally, if metrics
from multiple targets are compiled into one reference file, then it would
include redundant comments.

The "autogenerated" comment can still be enabled with a config option.

Signed-off-by: Anna Kapuscinska <anna@isovalent.com>